+++ lizard_fewsjdbc/admin.py
@@ -1,5 +1,16 @@
 from django.contrib import admin
-from lizard_fewsjdbc.models import IconStyle, JdbcSource, Threshold
+from lizard_fewsjdbc.models import (
+    IconStyle,
+    IconStyleWebRS,
+    JdbcSource,
+    Threshold,
+    WebRSSource,
+    FilterCache,
+    LocationCache,
+    ParameterCache,
+    TimeseriesCache,
+    FilterRootWebRSSource
+)
 
 
 class JdbcSourceAdmin(admin.ModelAdmin):
@@ -15,11 +26,35 @@ class IconStyleAdmin(admin.ModelAdmin):
         'fews_parameter', 'icon', 'color')
 
 
+class IconStyleWebRSAdmin(admin.ModelAdmin):
+    list_display = (
+        '__unicode__', 'jdbc_source', 'fews_filter', 'fews_location',
+        'fews_parameter', 'icon', 'color')
+    list_filter = (
+        'jdbc_source', 'fews_filter', 'fews_location',
+        'fews_parameter', 'icon', 'color')
+
+
 class ThresholdAdmin(admin.ModelAdmin):
     list_display = ('name', 'filter_id', 'parameter_id', 'location_id',
         'value')
 
 
+class WebRSSourceAdmin(admin.ModelAdmin):
+    list_display = ('source_code', 'version', 'base_path')
+
+
+class FilterRootWebRSSourceAdmin(admin.ModelAdmin):
+    list_display = ('name', 'slug', 'filter_tree_root', 'webrs_source')
+
+
 admin.site.register(JdbcSource, JdbcSourceAdmin)
 admin.site.register(IconStyle, IconStyleAdmin)
+admin.site.register(IconStyleWebRS, IconStyleWebRSAdmin)
 admin.site.register(Threshold, ThresholdAdmin)
+admin.site.register(WebRSSource, WebRSSourceAdmin)
+admin.site.register(FilterRootWebRSSource, FilterRootWebRSSourceAdmin)
+admin.site.register(FilterCache)
+admin.site.register(LocationCache)
+admin.site.register(ParameterCache)
+admin.site.register(TimeseriesCache)