+++ lizard_fewsjdbc/tasks.py
@@ -1,10 +1,13 @@
 import logging
 
 from celery.task import task
+from django.db import transaction
 
 from lizard_task.handler import get_handler
 from lizard_fewsjdbc.models import JdbcSource
 from lizard_fewsjdbc.models import CACHE_TIMEOUT
+from lizard_fewsjdbc.models import WebRSSource
+from lizard_fewsjdbc.cache_fews import FewsJDBCImporter
 
 
 @task
@@ -50,7 +53,7 @@ def load_parameters(timeout, jdbc_source,
 
 @task
 def rebuild_jdbc_cache_task(username=None, db_name=None,
-                       taskname=None, *args, **options):
+                            taskname=None, *args, **options):
     """
     Rebuild filter cache for fewsjdbc.
 
@@ -120,3 +123,40 @@ def rebuild_jdbc_cache(logger, *args, **options):
             logger.exception('Exception while syncing %s.' % jdbc_source)
     logger.info('Finished.')
     return 'OK'
+
+
+@transaction.commit_on_success
+def rebuild_restws_cache(logger, source_code=None):
+    """
+    Rebuild fews caches: (filters, locations, parameters, timeseries)
+
+    Raise exception to rollback transaction on error
+    """
+
+    if logger is None:
+        logger = logging.getLogger("tasks.rebuild_restws_cache")
+
+    importer = FewsJDBCImporter(logger)
+    importer.remove_all_caches()
+
+    if source_code is not None:
+        webrs_sources = [WebRSSource.objects.get(source_code=source_code)]
+    else:
+        webrs_sources = list(WebRSSource.objects.all())
+
+    for webrs_source in webrs_sources:
+        logger.info("Cache source {}".format(webrs_source.source_code))
+        ok = importer.cache_resources(webrs_source)
+        if ok is False:
+            raise Exception("Error on cache, rollback all changes, "
+                            "for info check previous logging")
+
+
+@task
+def rebuild_restws_cache_task(username=None, source_code=None,
+                              taskname=None, levelno=20):
+    """
+    Rebuild filter, location, parameter, timeseries cache for fewsjdbc.
+
+    """
+    rebuild_restws_cache(None, source_code)